fastapi_2/requirements.txt
databases==0.5.5
fastapi==0.74.1
uvicorn==0.17.5
asyncpg==0.25.0
pydantic==1.9.0
passlib==1.7.4
python-jose==3.3.0
SQLAlchemy==1.4.31